Move Stream/Enumeration under Stream/IsStream
